requirements.txt
asgiref==3.3.0
Django==3.1.3
djangorestframework==3.12.1
gunicorn==20.0.4
mysqlclient==2.0.1
pytz==2020.4
sqlparse==0.4.1
